fix(clauderon): change usage window types from i64 to f64 for typeshare compatibility
typeshare doesn't support i64 or u64 types. Changed to f64 which is
supported and semantically correct since TypeScript numbers are 64-bit
floats anyway.

Fixes CI error: 'Failed to parse a Rust type: Unsupported type: "i64"'

+++ packages/clauderon/src/api/claude_client.rs
@@ -29,16 +29,16 @@ struct ClaudeUsageResponse {
 #[derive(Debug, Deserialize)]
 struct UsageData {
     #[serde(default)]
-    current: i64,
+    current: f64,
     #[serde(default = "default_limit")]
-    limit: i64,
+    limit: f64,
     #[serde(default)]
     utilization: f64,
     resets_at: Option<String>,
 }
 
-fn default_limit() -> i64 {
-    100
+fn default_limit() -> f64 {
+    100.0
 }
 
 /// Client for Claude.ai API operations
